refactor: DriveFileManagerConstants is now in a standalone file
